+++ app/src/main/AndroidManifest.xml
@@ -3,7 +3,7 @@
     package="com.bitmovin.trialday">
 
     <application
-        android:name=".MainApplication"
+        android:name=".app.MainApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.TrialDay">
         <activity
-            android:name=".MainActivity"
+            android:name=".app.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
