+++ app/src/main/res/layout/activity_main.xml
@@ -2,12 +2,12 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".MainActivity">
+    tools:context=".app.MainActivity">
 
     <data>
         <variable
             name="viewModel"
-            type="com.bitmovin.trialday.MainViewModel" />
+            type="com.bitmovin.trialday.app.MainViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
